Renaming for clearer intent of classes
Refs #21251
