+++ qt/python/mantidqt/widgets/codeeditor/editor.py
@@ -24,103 +24,5 @@ from qtpy.QtWidgets import QStatusBar, QTabWidget, QVBoxLayout, QWidget
 from mantidqt.utils.qt import import_qtlib
 from .execution import PythonCodeExecution
 
-DEFAULT_EDITOR_LANGUAGE = "Python"
-
-EXECUTION_CLS = PythonCodeExecution
-
-IDLE_STATUS_MSG = "Status: Idle"
-
-RUNNING_STATUS_MSG = "Status: Running"
-
 # Import single-file editor from C++ wrapping
 CodeEditor = import_qtlib('_widgetscore', 'mantidqt.widgets', 'ScriptEditor')
-
-
-class ExecutableCodeEditor(QWidget):
-
-    def __init__(self, language=DEFAULT_EDITOR_LANGUAGE, parent=None):
-        """
-
-        :param language: Language for syntax highlighting
-        :param user_globals: Dictionary for global context of execution.
-        :param user_locals: Dictionary for local context of execution
-        :param parent: A parent QWidget
-        """
-        super(ExecutableCodeEditor, self).__init__(parent)
-
-        # layout
-        self.editor = CodeEditor(language, self)
-        self.status = QStatusBar(self)
-        layout = QVBoxLayout()
-        layout.addWidget(self.editor)
-        layout.addWidget(self.status)
-        self.setLayout(layout)
-        layout.setContentsMargins(0, 0, 0, 0)
-
-        # presenter
-        self._presenter = ExecutableCodeEditorPresenter(self, PythonCodeExecution())
-
-    def execute_all_async(self):
-        self._presenter.req_execute_all_async()
-
-    def set_editor_readonly(self, ro):
-        self.editor.setReadOnly(ro)
-
-    def set_status_message(self, msg):
-        self.status.showMessage(msg)
-
-
-class ExecutableCodeEditorPresenter(QObject):
-    """Presenter part of MVP to control actions on the editor"""
-
-    def __init__(self, view, model):
-        super(ExecutableCodeEditorPresenter, self).__init__()
-        self.view = view
-        self.model = model
-
-        # connect signals
-        self.model.sig_exec_success.connect(self.on_exec_success)
-        self.model.sig_exec_error.connect(self.on_exec_error)
-
-        # starts idle
-        self.view.set_status_message(IDLE_STATUS_MSG)
-
-    def req_execute_all_async(self):
-        text = self.view.editor.text()
-        if not text:
-            return
-        self.view.set_editor_readonly(True)
-        self.view.set_status_message(RUNNING_STATUS_MSG)
-        return self.model.execute_async(text)
-
-    def on_exec_success(self):
-        self.view.set_status_message(IDLE_STATUS_MSG)
-
-    def on_exec_error(self):
-        self.view.set_status_message(IDLE_STATUS_MSG)
-
-
-class MultiFileCodeEditor(QWidget):
-    """Provides a tabbed widget for editing multiple files"""
-
-    def __init__(self, parent=None):
-        super(MultiFileCodeEditor, self).__init__(parent)
-
-        # layout
-        self.editors = QTabWidget(self)
-        self.editors.setMovable(True)
-        layout = QVBoxLayout()
-        layout.addWidget(self.editors)
-        self.setLayout(layout)
-        layout.setContentsMargins(0, 0, 0, 0)
-
-        # add a single editor by default
-        self.append_new_editor(DEFAULT_EDITOR_LANGUAGE)
-
-    @property
-    def editor_count(self):
-        return self.editors.count()
-
-    def append_new_editor(self, language):
-        title = "New"
-        self.editors.addTab(ExecutableCodeEditor(language, self.editors), title)
